+++ release-notes.md
@@ -7,6 +7,8 @@
 * Added support for running it from the game folder (thanks to collaboration with Platonymous!) or another app.
 * Added file count and unpack time to log.
 * Changed map format from `.tbin` to `.tmx` (thanks to Platonymous!).
+* Simplified usage on Linux/MacOS with launcher scripts.
+* Fixed error unpacking files on MacOS.
 * Fixed BMFont file extension set to `.xml` instead of `.fnt` (thanks to Platonymous!).
 * Fixed unsupported XNB files not always copied into the export folder.